issue10841 add convertion to internal types for DynamicModelSerializer
